requirements.txt

langchain==0.2.*
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
mysql-connector-python
python-dotenv
